requirements.txt
# My-LM — pinned dependencies known to work together.
#
# IMPORTANT: PyTorch must be installed separately from the CUDA index URL.
# The setup script handles this for you:
#
#   pip install torch==2.5.1 torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cu121
#
# Do NOT install `xformers` — the available wheels are pinned to PyTorch 2.10+
# and break SDXL on 2.5.x. Diffusers falls back to PyTorch's built-in attention.

# ── Core model loading ────────────────────────────────────────────────────
transformers>=4.57.6
accelerate==1.13.0
safetensors==0.7.0
numpy==2.4.3
pillow==12.1.1

# ── Diffusion + prompt conditioning ───────────────────────────────────────
diffusers==0.37.1
compel==2.3.1                    # long-prompt support; image_bridge.py
                                 # monkey-patches EmbeddingsProviderMulti
                                 # for the `empty_z` SDXL bug

# ── QLoRA training ────────────────────────────────────────────────────────
peft==0.18.1
bitsandbytes==0.49.2             # 4-bit quantization (Windows wheel works)
datasets==4.8.4

# ── Face detector (ADetailer-style face fix) ─────────────────────────────
ultralytics==8.4.36
opencv-python==4.13.0.92

# ── ESRGAN upscalers ──────────────────────────────────────────────────────
spandrel==0.4.2
spandrel_extra_arches==0.2.0

# ── Hugging Face hub & fast downloads ─────────────────────────────────────
huggingface_hub==0.36.2
hf_transfer==0.1.9

# ── BookMind RAG (optional but bundled by default) ────────────────────────
pymongo>=4.6
python-dotenv>=1.0
sentence-transformers>=2.7
requests>=2.31
beautifulsoup4>=4.12
